ai-mcp/pyproject.toml
[project]
name = "ai-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "bilibili-api-python>=17.3.0",
    "mcp[cli]>=1.9.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "pdfplumber>=0.11.7",
    "psutil>=7.0.0",
]
